Files renamed to be more precise, typo in comment fixed.
