+++ monkey/monkey_island/cc/services/attack/__init__.py
